+++ src/nationalarchives/components/tabs/tabs.mjs
@@ -115,7 +115,7 @@ export class Tabs {
         $tabItem.removeAttribute("hidden");
         $tabItem.setAttribute("tabindex", "0");
       } else {
-        $tabItem.setAttribute("hidden", true);
+        $tabItem.setAttribute("hidden", "");
         $tabItem.setAttribute("tabindex", "-1");
       }
     });