+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 ### Removed
 ### Fixed
+
+- Added missing `type` attribute on accordion buttons
+
 ### Security
 
 ## [0.8.0](https://github.com/nationalarchives/tna-frontend/compare/v0.7.2...v0.8.0) - 2025-01-15