+++ CHANGELOG.md
@@ -5,18 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/nationalarchives/tna-frontend/compare/v0.8.0...HEAD)
+## [Unreleased](https://github.com/nationalarchives/tna-frontend/compare/v0.8.1...HEAD)
 
 ### Added
 ### Changed
 ### Deprecated
 ### Removed
 ### Fixed
+### Security
+
+## [0.8.1](https://github.com/nationalarchives/tna-frontend/compare/v0.8.0...v0.8.1) - 2025-01-21
+
+### Fixed
 
 - Added missing `type` attribute on accordion buttons
 - Fixed issue with invalid HTML in accordion JavaScript - headings are not permitted inside `<button>` elements
-
-### Security
 
 ## [0.8.0](https://github.com/nationalarchives/tna-frontend/compare/v0.7.2...v0.8.0) - 2025-01-15